refactor: move api folder to outside
